+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 #### Features
 * Added a shopping list menu that can be dragged around.
 
+## [Unreleased]
+
+#### Features
+
+* Added mod badge and config menu.
+* Added an option to toggle whether card arts are used to display cards in the shopping list.